+++ Source/Testably.Expectations/Core/Nodes/AndNode.cs
@@ -1,4 +1,6 @@
-﻿using System.Threading;
+﻿using System;
+using System.Collections.Generic;
+using System.Threading;
 using System.Threading.Tasks;
 using Testably.Expectations.Core.Constraints;
 using Testably.Expectations.Core.EvaluationContext;
@@ -6,71 +8,105 @@ using Testably.Expectations.Core.Helpers;
 
 namespace Testably.Expectations.Core.Nodes;
 
-internal class AndNode : CombinationNode
+internal class AndNode : Node
 {
-	public override Node Left { get; }
-	public override Node Right { get; set; }
-	private readonly string _textSeparator;
+	private const string DefaultSeparator = " and ";
+	private Node Current { get; set; }
+	private string? _currentSeparator;
+	private readonly List<(string, Node)> _nodes = new();
 
-	public AndNode(Node left, Node right, string textSeparator = " and ")
+	public AndNode(Node node)
 	{
-		Left = left;
-		Right = right;
-		_textSeparator = textSeparator;
+		Current = node;
 	}
 
 	/// <inheritdoc />
-	public override async Task<ConstraintResult> IsMetBy<TValue>(
-		TValue? value,
-		IEvaluationContext context,
-		CancellationToken cancellationToken)
-		where TValue : default
+	public override void AddConstraint(IConstraint constraint)
+		=> Current.AddConstraint(constraint);
+
+	/// <inheritdoc />
+	public override void AddMapping<TValue, TTarget>(
+		IValueConstraint<TValue>? precondition,
+		PropertyAccessor<TValue, TTarget?> propertyAccessor,
+		Func<PropertyAccessor, string, string>? expectationTextGenerator = null)
+		where TTarget : default
+		=> Current.AddMapping(precondition, propertyAccessor, expectationTextGenerator);
+
+	public override void AddNode(Node node, string? separator = null)
 	{
-		ConstraintResult leftResult = await Left.IsMetBy(value, context, cancellationToken);
-		if (leftResult.IgnoreFurtherProcessing)
+		_nodes.Add((_currentSeparator ?? DefaultSeparator, Current));
+		Current = node;
+		_currentSeparator = separator;
+	}
+
+	/// <inheritdoc />
+	public override async Task<ConstraintResult> IsMetBy<TValue>(TValue? value,
+		IEvaluationContext context,
+		CancellationToken cancellationToken) where TValue : default
+	{
+		_nodes.Add((_currentSeparator ?? DefaultSeparator, Current));
+		ConstraintResult? combinedResult = null;
+		foreach ((string separator, Node node) in _nodes)
 		{
-			return leftResult;
+			if (node is ExpectationNode expectationNode && expectationNode.IsEmpty())
+			{
+				continue;
+			}
+
+			ConstraintResult result = await node.IsMetBy(value, context, cancellationToken);
+			combinedResult = CombineResults(combinedResult, result, separator);
+			if (result.IgnoreFurtherProcessing)
+			{
+				return combinedResult;
+			}
 		}
 
-		ConstraintResult rightResult = await Right.IsMetBy(value, context, cancellationToken);
+		return combinedResult!;
+	}
+
+	/// <inheritdoc />
+	public override void SetReason(BecauseReason becauseReason)
+		=> Current.SetReason(becauseReason);
+
+	/// <inheritdoc />
+	public override string ToString()
+		=> string.Join(" and ", _nodes) + " and " + Current;
+
+	private ConstraintResult CombineResults(ConstraintResult? combinedResult,
+		ConstraintResult result, string separator)
+	{
+		if (combinedResult == null)
+		{
+			return result;
+		}
 
 		string combinedExpectation =
-			$"{leftResult.ExpectationText}{_textSeparator}{rightResult.ExpectationText}";
+			$"{combinedResult.ExpectationText}{separator}{result.ExpectationText}";
 
-		if (leftResult is ConstraintResult.Failure leftFailure &&
-		    rightResult is ConstraintResult.Failure rightFailure)
+		if (combinedResult is ConstraintResult.Failure leftFailure &&
+		    result is ConstraintResult.Failure rightFailure)
 		{
 			return leftFailure.CombineWith(
 				combinedExpectation,
 				CombineResultTexts(leftFailure.ResultText, rightFailure.ResultText));
 		}
 
-		if (leftResult is ConstraintResult.Failure onlyLeftFailure)
+		if (combinedResult is ConstraintResult.Failure onlyLeftFailure)
 		{
 			return onlyLeftFailure.CombineWith(
 				combinedExpectation,
 				onlyLeftFailure.ResultText);
 		}
 
-		if (rightResult is ConstraintResult.Failure onlyRightFailure)
+		if (result is ConstraintResult.Failure onlyRightFailure)
 		{
 			return onlyRightFailure.CombineWith(
 				combinedExpectation,
 				onlyRightFailure.ResultText);
 		}
 
-		return leftResult.CombineWith(combinedExpectation, "");
+		return combinedResult.CombineWith(combinedExpectation, "");
 	}
-
-	/// <inheritdoc />
-	public override void SetReason(BecauseReason reason)
-	{
-		Right.SetReason(reason);
-	}
-
-	/// <inheritdoc />
-	public override string ToString()
-		=> $"({Left} AND {Right})";
 
 	private static string CombineResultTexts(string leftResultText, string rightResultText)
 	{
