--- Source/Testably.Expectations/Core/Nodes/CombinationNode.cs
@@ -1,7 +1,0 @@
-﻿namespace Testably.Expectations.Core.Nodes;
-
-internal abstract class CombinationNode : Node
-{
-	public abstract Node Left { get; }
-	public abstract Node Right { get; set; }
-}
