--- Source/Testably.Expectations/Core/Nodes/ManipulationNode.cs
@@ -1,6 +1,0 @@
-﻿namespace Testably.Expectations.Core.Nodes;
-
-internal abstract class ManipulationNode : Node
-{
-	public abstract Node Inner { get; set; }
-}
