+++ Source/Testably.Expectations/Core/Nodes/WhichNode.cs
@@ -3,64 +3,64 @@ using System.Threading;
 using System.Threading.Tasks;
 using Testably.Expectations.Core.Constraints;
 using Testably.Expectations.Core.EvaluationContext;
+using Testably.Expectations.Core.Helpers;
 
 namespace Testably.Expectations.Core.Nodes;
 
-internal class WhichNode<TSource, TProperty> : ManipulationNode
+internal class WhichNode<TSource, TProperty> : Node
 {
-	public override Node Inner { get; set; }
-	private readonly Func<PropertyAccessor, string, string> _expectationTextGenerator;
-	private readonly PropertyAccessor _propertyAccessor;
+	private Node? _inner;
+	private readonly Func<TSource, TProperty?> _propertyAccessor;
 
-	public WhichNode(PropertyAccessor propertyAccessor, Node inner,
-		Func<PropertyAccessor, string, string>? expectationTextGenerator = null)
+	public WhichNode(
+		Func<TSource, TProperty?> propertyAccessor)
 	{
 		_propertyAccessor = propertyAccessor;
-		_expectationTextGenerator = expectationTextGenerator ?? DefaultExpectationTextGenerator;
-		Inner = inner;
 	}
 
 	/// <inheritdoc />
-	public override async Task<ConstraintResult> IsMetBy<TValue>(
+	public override void AddConstraint(IConstraint constraint)
+		=> _inner?.AddConstraint(constraint);
+
+	/// <inheritdoc />
+	public override void AddMapping<TValue, TTarget>(IValueConstraint<TValue>? precondition,
+		PropertyAccessor<TValue, TTarget?> propertyAccessor,
+		Func<PropertyAccessor, string, string>? expectationTextGenerator = null)
+		where TTarget : default
+		=> _inner?.AddMapping(precondition, propertyAccessor, expectationTextGenerator);
+
+	/// <inheritdoc />
+	public override void AddNode(Node node, string? separator = null)
+	{
+		_inner = node;
+	}
+
+	/// <inheritdoc />
+	public override Task<ConstraintResult> IsMetBy<TValue>(
 		TValue? value,
 		IEvaluationContext context,
-		CancellationToken cancellationToken)
-		where TValue : default
+		CancellationToken cancellationToken) where TValue : default
 	{
-		if (_propertyAccessor is PropertyAccessor<TSource, TProperty> propertyAccessor)
+		if (value is not TSource typedValue)
 		{
-			if (value is not TSource typedValue)
-			{
-				throw new InvalidOperationException(
-					$"The property type for the actual value in the which node did not match.{Environment.NewLine}Expected {typeof(TSource).Name},{Environment.NewLine}but found {value?.GetType().Name}");
-			}
-
-			if (propertyAccessor.TryAccessProperty(
-				typedValue,
-				out TProperty? matchingValue))
-			{
-				ConstraintResult result =
-					(await Inner.IsMetBy(matchingValue, context, cancellationToken))
-					.UpdateExpectationText(r
-						=> _expectationTextGenerator(_propertyAccessor, r.ExpectationText));
-				return result.UseValue(value);
-			}
-
 			throw new InvalidOperationException(
-				$"The property type for the which node did not match.{Environment.NewLine}Expected {typeof(TProperty).Name},{Environment.NewLine}but found {matchingValue?.GetType().Name}");
+				$"The property type for the actual value in the which node did not match.{Environment.NewLine}Expected {typeof(TValue).Name},{Environment.NewLine}but found {value?.GetType().Name}");
 		}
 
-		throw new InvalidOperationException(
-			$"The property accessor for the which node did not match.{Environment.NewLine}Expected {typeof(PropertyAccessor<TValue, TProperty>).FullName},{Environment.NewLine}but found {_propertyAccessor.GetType().FullName}");
+		if (_inner == null)
+		{
+			throw new InvalidOperationException("No inner node specified for the which node.");
+		}
+
+		TProperty? matchingValue = _propertyAccessor(typedValue);
+		return _inner.IsMetBy(matchingValue, context, cancellationToken);
 	}
+
+	/// <inheritdoc />
+	public override void SetReason(BecauseReason becauseReason)
+		=> _inner?.SetReason(becauseReason);
 
 	/// <inheritdoc />
 	public override string ToString()
-		=> $".{_propertyAccessor} {Inner}";
-
-	private string DefaultExpectationTextGenerator(PropertyAccessor propertyAccessor,
-		string expectationText)
-	{
-		return propertyAccessor + expectationText;
-	}
+		=> _propertyAccessor + base.ToString();
 }
